fix: Tree shaking should works on withContext
