+++ components/config-provider/context.tsx
@@ -1,0 +1,78 @@
+import * as React from 'react';
+import createReactContext from '@ant-design/create-react-context';
+import defaultRenderEmpty, { RenderEmptyHandler } from './renderEmpty';
+import { Locale } from '../locale-provider';
+
+export interface CSPConfig {
+  nonce?: string;
+}
+
+export interface ConfigConsumerProps {
+  getPopupContainer?: (triggerNode: HTMLElement) => HTMLElement;
+  rootPrefixCls?: string;
+  getPrefixCls: (suffixCls: string, customizePrefixCls?: string) => string;
+  renderEmpty: RenderEmptyHandler;
+  csp?: CSPConfig;
+  autoInsertSpaceInButton?: boolean;
+  locale?: Locale;
+  pageHeader?: {
+    ghost: boolean;
+  };
+}
+
+export const ConfigContext = createReactContext<ConfigConsumerProps>({
+  // We provide a default function for Context without provider
+  getPrefixCls: (suffixCls: string, customizePrefixCls?: string) => {
+    if (customizePrefixCls) return customizePrefixCls;
+
+    return `ant-${suffixCls}`;
+  },
+
+  renderEmpty: defaultRenderEmpty,
+});
+
+export const ConfigConsumer = ConfigContext.Consumer;
+
+// =========================== withConfigConsumer ===========================
+// We need define many types here. So let's put in the block region
+type IReactComponent<P = any> =
+  | React.StatelessComponent<P>
+  | React.ComponentClass<P>
+  | React.ClassicComponentClass<P>;
+
+interface BasicExportProps {
+  prefixCls?: string;
+}
+
+interface ConsumerConfig {
+  prefixCls: string;
+}
+
+interface ConstructorProps {
+  displayName?: string;
+}
+export function withConfigConsumer<ExportProps extends BasicExportProps>(config: ConsumerConfig) {
+  return function withConfigConsumerFunc<ComponentDef>(
+    Component: IReactComponent,
+  ): React.SFC<ExportProps> & ComponentDef {
+    // Wrap with ConfigConsumer. Since we need compatible with react 15, be care when using ref methods
+    const SFC = ((props: ExportProps) => (
+      <ConfigConsumer>
+        {(configProps: ConfigConsumerProps) => {
+          const { prefixCls: basicPrefixCls } = config;
+          const { getPrefixCls } = configProps;
+          const { prefixCls: customizePrefixCls } = props;
+          const prefixCls = getPrefixCls(basicPrefixCls, customizePrefixCls);
+          return <Component {...configProps} {...props} prefixCls={prefixCls} />;
+        }}
+      </ConfigConsumer>
+    )) as React.SFC<ExportProps> & ComponentDef;
+
+    const cons: ConstructorProps = Component.constructor as ConstructorProps;
+    const name = (cons && cons.displayName) || Component.name || 'Component';
+
+    SFC.displayName = `withConfigConsumer(${name})`;
+
+    return SFC;
+  };
+}
