+++ components/typography/Base.tsx
@@ -6,7 +6,8 @@ import findDOMNode from 'rc-util/lib/Dom/findDOMNode';
 import copy from 'copy-to-clipboard';
 import omit from 'omit.js';
 import ResizeObserver from 'rc-resize-observer';
-import { withConfigConsumer, ConfigConsumerProps, configConsumerProps } from '../config-provider';
+import { ConfigConsumerProps, configConsumerProps } from '../config-provider';
+import { withConfigConsumer } from '../config-provider/context';
 import LocaleReceiver from '../locale-provider/LocaleReceiver';
 import warning from '../_util/warning';
 import TransButton from '../_util/transButton';
